pyproject.toml
[tool.poetry]
name = "ditlep"
version = "0.1.0"
description = ""
authors = ["Marcuth <example@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.1"
dragon-city-utils = "^2.0.3"
pycryptodome = "^3.20.0"
httpx = "^0.26.0"


[tool.poetry.group.dev.dependencies]
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
